+++ src/noisegeneration.jl
@@ -89,6 +89,7 @@ end
     - the sampling frequency (in Hz)
     - an array containing the RMS of the temperature (in K) for each polarimeter, σ_k[i]= tsys_k[i] / sqrt(β_hz[i] * τ_s)
     - an array containing the knee frequency (in Hz) for each polarimeter
+    - an array containing the oof slope for each polarimeter
     - the MPI communicator
     - the random generator seed, as optional argument.
 
@@ -109,7 +110,7 @@ end
 
 
 """
-function  generate_noise_mpi(chunks, baselines_per_process, baseline_length_s, total_time, fsamp_hz, σ_k, fknee_hz, rank, comm, input_seed = missing)
+function  generate_noise_mpi(chunks, baselines_per_process, baseline_length_s, total_time, fsamp_hz, σ_k, fknee_hz, slope, rank, comm, input_seed = missing)
     
     total_number_of_polarimeters = length(σ_k)
     commsize = length(chunks)
@@ -130,7 +131,7 @@ function  generate_noise_mpi(chunks, baselines_per_process, baseline_length_s, t
         noise = Array{Float64}(undef, baselines_per_process[rank+1]*baseline_length_s*fsamp_hz)    
         
         if  0 <= rank < total_number_of_polarimeters  #if first "total_number_of_polarimeters" ranks
-            rng = CorrNoise.OofRNG(Random.MersenneTwister(seeds[pol_number]), -1, 1.15e-5, fknee_hz[pol_number], fsamp_hz)
+            rng = CorrNoise.OofRNG(Random.MersenneTwister(seeds[pol_number]), slope[pol_number], 1.15e-5, fknee_hz[pol_number], fsamp_hz)
             samples_per_pol = total_time*fsamp_hz
             pol_noise = Float64[CorrNoise.randoof(rng) * σ_k[pol_number] for i in 1:(samples_per_pol)]
         end
@@ -192,7 +193,7 @@ function  generate_noise_mpi(chunks, baselines_per_process, baseline_length_s, t
             previous_detector = chunks[1][1].pol_number
             noise = Float64[]  #noise rank 0
 
-            rng = CorrNoise.OofRNG(Random.MersenneTwister(seeds[1]), -1, 1.15e-5, fknee_hz[1], fsamp_hz)
+            rng = CorrNoise.OofRNG(Random.MersenneTwister(seeds[1]), slope[1], 1.15e-5, fknee_hz[1], fsamp_hz)
 
             for i in 1:length(chunks)   #loop on ranks
                 num_noise_samples = baselines_per_process[i]*baseline_length_s*fsamp_hz
@@ -201,7 +202,7 @@ function  generate_noise_mpi(chunks, baselines_per_process, baseline_length_s, t
                     cur_detector = chunks[i][j].pol_number
 
                     if cur_detector != previous_detector  #if new detector generate new noise
-                        rng = CorrNoise.OofRNG(Random.MersenneTwister(seeds[cur_detector]), -1, 1.15e-5, fknee_hz[cur_detector], fsamp_hz)
+                        rng = CorrNoise.OofRNG(Random.MersenneTwister(seeds[cur_detector]), slope[cur_detector], 1.15e-5, fknee_hz[cur_detector], fsamp_hz)
                     end
                     cur_num_noise_samples = chunks[i][j].num_of_elements*baseline_length_s*fsamp_hz
                     cur_noise = Float64[CorrNoise.randoof(rng) * σ_k[cur_detector] for i in 1:(cur_num_noise_samples)]