+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add functions to simulate ADCs [#24](https://github.com/lspestrip/Stripeline.jl/pull/24)
 - Add a high-level API for accessing the instrument database [#30](https://github.com/lspestrip/Stripeline.jl/pull/30)
 - Fix bandshapes for W-band polarimeters [#32](https://github.com/lspestrip/Stripeline.jl/pull/32)
+- Properly use the slope of the 1/f noise when generating maps from TODs [#31](https://github.com/lspestrip/Stripeline.jl/pull/31)
 
 # Version 0.4